+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Legenda for the issue kind:
 - [#559](https://github.com/ariadne-cps/ariadne/issues/559) (N) Introduce an optional ariadne_main.hpp that offers an ariadne_main() function handling all header tasks for an executable   
 - [#562](https://github.com/ariadne-cps/ariadne/issues/562) (N) Introduce a command line interface class for acquiring CLI input arguments setting logger verbosity, theme and scheduler, along with concurrency, graphics backend and drawer  
 - [#599](https://github.com/ariadne-cps/ariadne/issues/599) (A) Add extra Python bindings for HybridEnclosure, add Python subscripting for ListSet<HybridEnclosure> 
+- [#602](https://github.com/ariadne-cps/ariadne/issues/602) (A) Add extra Python bindings for Enclosure, add Python bindings for Point2d for graphics
 - [#551](https://github.com/ariadne-cps/ariadne/issues/551) (C) Make VectorFieldEvolver process sets in parallel (when splitting initially or during evolution)
 - [#556](https://github.com/ariadne-cps/ariadne/issues/556) (C) Disallow unsafe default value of upper semantics for orbit methods
 - [#565](https://github.com/ariadne-cps/ariadne/issues/565) (C) Draw lists of (Labelled/Hybrid)Enclosure in parallel